Refacotr protocol
