+++ sgl-router/src/protocols/generate/mod.rs
@@ -1,0 +1,8 @@
+// SGLang native Generate API module (/generate)
+
+pub mod request;
+pub mod types;
+
+// Re-export main types for convenience
+pub use request::GenerateRequest;
+pub use types::{GenerateParameters, InputIds, SamplingParams};
